Boundary conditions for space requirements were wrong; we need to be able to hold at least two keys per sector for a BTree (for the "not full" invariant) So, added helper methods to calculate this requirement.
Deciding on the mid. point must consider that we may need to reserve more
than half of a sector for free space (on badly tuned parameters). If not, we
can crash because we start scanning non-existant records.

The boundary conditions in traverse_down must be juuust right. They were off
by one, and this broke stuff.


git-svn-id: https://svn.code.sf.net/p/lurker/code/trunk@457 44f0b718-a8ad-45a4-bab0-870672037115